Prevented Workbench Freezeup whenever Analysis plug-in is resized (Mac-specific)
The freeze up occurs when a chart is drawn.

Made it so when resizing starts, the current chart is destroyed. Then
resizing takes place. After resizing has completed, the chart is
redrawn. 

This is a quick fix on what appears to be an underlying issue with the
JFreeChart plug-in on Mac.